Use u64 instead of u32 for LineId
We are no longer limited to 2^32 unique lines and more importantly we will be
able to encode additional information to the LineId, which will be useful in
future optimization.

Tested with 44247 patches of SUSE's kernel-source branch SLE15-SP1, the time to
apply using 8 threads increased by +0.8% and memory consumption increased by
+5%. This is acceptable in my opinion.

+++ src/libpatch/line_interner.rs
@@ -18,7 +18,7 @@ impl fmt::Display for Stats {
 
 /// ID that is given to every unique line
 #[derive(Clone, Copy, Eq, Hash, PartialEq)]
-pub struct LineId(u32); // 4G interned slices ought to be enough for everybody...
+pub struct LineId(u64);
 
 impl<'a> fmt::Debug for LineId {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
@@ -52,7 +52,7 @@ impl<'a> LineInterner<'a> {
         // otherwise it would not exist.
         debug_assert!(!bytes.is_empty());
 
-        LineId(self.set.insert_full(bytes).0 as u32)
+        LineId(self.set.insert_full(bytes).0 as u64)
     }
 
     /// Get the line for given `LineId`. Returns `None` if that id isn't known.
